# api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Sphere Swagger",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/api/expectation/create": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new expectation.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expectations"
                ],
                "summary": "ADMIN",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Title of the expectation",
                        "name": "title",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expectation created successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict occurred",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    }
                }
            }
        },
        "/api/expectation/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete an expectation by ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expectations"
                ],
                "summary": "ADMIN",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Expectation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expectation deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict occurred",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    }
                }
            }
        },
        "/api/expectation/update/{id}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update an existing expectation by ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expectations"
                ],
                "summary": "ADMIN",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Expectation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Title of the expectation",
                        "name": "title",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expectation updated successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict occurred",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    }
                }
            }
        },
...
        "/api/lead/create": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new lead with the given title.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leads"
                ],
                "summary": "ADMIN",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Title of the lead",
                        "name": "title",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lead successfully created",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "409": {
                        "description": "Lead creation conflict",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    }
                }
            }
        },
        "/api/lead/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete a lead by ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leads"
                ],
                "summary": "ADMIN",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lead ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lead deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict occurred",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    }
                }
            }
        },
        "/api/lead/get-all/{companyId}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update the data associated with a lead",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leads"
                ],
                "summary": "ALL",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.GetLeadListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    }
                }
            }
        },
        "/api/lead/get-lead-common": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a common lead by ID and type.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leads"
                ],
                "summary": "ADMIN",
                "parameters": [
                    {
                        "description": "Lead ID",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.GetLeadCommonRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lead details retrieved",
                        "schema": {
                            "$ref": "#/definitions/pb.GetLeadCommonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    }
                }
            }
        },
        "/api/lead/get-lead-reports": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieves lead reports for a given date range.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leads"
                ],
                "summary": "CEO , ADMIN",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date in YYYY-MM-DD format",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date in YYYY-MM-DD format",
                        "name": "till",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lead reports response",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict error with details",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    }
                }
            }
        },
        "/api/lead/update/{id}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update an existing lead by ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leads"
                ],
                "summary": "ADMIN",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lead ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Title of the lead",
                        "name": "title",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lead updated successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict occurred",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    }
                }
            }
        },
        "/api/leadData/change-lead-data": {
            "patch": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update the data associated with a lead",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leadData"
                ],
                "summary": "Change lead data",
                "parameters": [
                    {
                        "description": "Lead change request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.ChangeLeadPlaceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pb.ChangeLeadPlaceRequest": {
            "type": "object",
            "properties": {
                "changedSet": {
                    "$ref": "#/definitions/pb.ChangeLeadDataRequest"
                },
                "leadDataId": {
                    "type": "string"
                }
            }
        },
...
        "pb.DynamicSection": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "pb.GetLeadCommonRequest": {
            "type": "object",
            "properties": {
                "requests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.LeadCommonRequest"
                    }
                }
            }
        },
        "pb.GetLeadCommonResponse": {
            "type": "object",
            "properties": {
                "expectations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.Section"
                    }
                },
                "leads": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.Section"
                    }
                },
                "sets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.Section"
                    }
                }
            }
        },
        "pb.GetLeadListResponse": {
            "type": "object",
            "properties": {
                "sections": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.DynamicSection"
                    }
                }
            }
        },
...
        "pb.LeadCommonRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "utils.AbsResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}